[AF-463] feat (FRONT topics): refactor component and service arbo
