+++ front/src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './features/auth/components/home/home.component';
 import { UnauthGuard } from './guards/unauth.guard';
 import { AuthGuard } from './guards/auth.guard';
-import { UserDetailsComponent } from './components/user-details/user-details.component';
+import { UserDetailsComponent } from './features/feed/components/user-details/user-details.component';
 
 const routes: Routes = [
   {
@@ -18,11 +18,6 @@ const routes: Routes = [
     loadChildren: () =>
       import('./features/feed/feed.module').then((m) => m.FeedModule),
   },
-  {
-      title: 'UserDetails',
-      path: 'user-details',
-      component: UserDetailsComponent,
-    },
 ];
 
 @NgModule({
