+++ front/src/app/app.module.ts
@@ -22,7 +22,7 @@ import { LoginComponent } from './features/auth/components/login/login.component
 import { RegisterComponent } from './features/auth/components/register/register.component';
 import { JwtInterceptor } from './interceptors/jwt.interceptor';
 import { FeedComponent } from './features/feed/components/feed/feed.component';
-import { UserDetailsComponent } from './components/user-details/user-details.component';
+import { UserDetailsComponent } from './features/feed/components/user-details/user-details.component';
 import { TopicsComponent } from './features/feed/components/topics/topics.component';
 
 const materialModule = [
