+++ front/src/app/features/feed/components/topics/topics.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { TopicService } from '../../services/topic.service';
-import { UserService } from 'src/app/services/user.service';
 import { SessionService } from 'src/app/services/session.service';
-import { User } from 'src/app/interfaces/user.interface';
+import { User } from 'src/app/features/feed/interfaces/user.interface';
+import { UserService } from '../../services/user.service';
 
 @Component({
   selector: 'app-topics',
