+++ front/src/app/features/feed/components/user-details/user-details.component.spec.ts
