+++ front/src/app/features/feed/interfaces/user.interface.ts
